requirements.txt
django==3.1.1
djangorestframework==3.11.1
pyjwt==1.7.1
pyyaml==5.3.1